+++ lib/types.ts
@@ -80,6 +80,9 @@ export interface SequenceItem {
     /** Tests that the AudioPlayer clears the queue with the given clear behavior. */
     clearsQueue? : string;
 
+    /** Tests that the VideoPlayer is used to play a stream. */
+    playsVideo? : PlayVideoConfig;
+
     /** Any additional fields for custom validators */
     [key : string] : any;
 }
@@ -90,6 +93,12 @@ export interface PlayStreamConfig {
     previousToken? : string;
     url? : string;
     offset? : number;
+}
+
+export interface PlayVideoConfig {
+    source? : string;
+    titel? : string;
+    subtitle? : string;
 }
 
 export abstract class ResponseValidator {