+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## current master
 * add supportedInterfaces to request builder
 * add missing feature `withSessionAttributes` (#3)
+* add validator for video playback
 
 ## 2.0.1
 * adding more supported types to attributes check